+++ pol-core/bscript/executor.h
@@ -97,7 +97,6 @@ public:
   bool error_;
   bool halt_;
   bool run_ok_;
-  bool can_access_offline_mobiles_;
 
   virtual ExecutorType type() { return ExecutorType::EXECUTOR;  }
