+++ pol-core/pol/module/partymod.h
@@ -26,7 +26,6 @@ class Party;
 }
 namespace Module
 {
-extern Bscript::BApplicObjType party_type;
 
 class PartyExecutorModule : public Bscript::TmplExecutorModule<PartyExecutorModule, Core::PolModule>
 {
